% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupaggs.R
\name{group.aggs}
\alias{group.aggs}
\title{filter rows based on a column compared to a statistic within a group}
\usage{
group.aggs(data, group.cols, agg.funct, agg.col, symbol)
}
\arguments{
\item{data}{dataset, must be a data.table}

\item{group.cols}{columns to group by}

\item{agg.funct}{function which value is calculated for the group}

\item{agg.col}{numerical column which value is compared to the functions value}

\item{symbol}{symbol of the comparison}
}
\value{
a filtered data.table object
}
\description{
filter rows based on a column compared to a statistic within a group
}
\examples{
#this returns rows of the iris set that have their Sepal.Length higher than the mean of their Species
iris_dt <- as.data.table(iris)
group.aggs(iris_dt,'Species',mean,'Sepal.Length','>')
}
