% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuncts-package.R
\docType{package}
\name{helperfuncts-package}
\alias{helperfuncts}
\alias{helperfuncts-package}
\title{helperfuncts: Useful Function Wrappers for data.table}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
