% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuncts-package.R
\docType{package}
\name{helperfuncts-package}
\alias{helperfuncts}
\alias{helperfuncts-package}
\title{helperfuncts: Useful Function Wrappers for data.table}
\description{
Useful wrappers for data.table
}
\author{
\strong{Maintainer}: Piotr Sternicki \email{sternicki.p@gmail.com}

}
\keyword{internal}
